Reorg to change height finder to terrain query class
